[file/#122] : 경로수정
